+++ src/shared/types/material.hpp
@@ -10,24 +10,24 @@
 namespace fr {
 
 struct Material {
-    explicit Material(uint64_t id);
+    explicit Material(uint32_t id);
 
-    explicit Material(uint64_t id, uint64_t shader);
+    explicit Material(uint32_t id, uint32_t shader);
 
-    explicit Material(uint64_t id, uint64_t shader, bool emissive);
+    explicit Material(uint32_t id, uint32_t shader, bool emissive);
 
     // FOR MSGPACK ONLY!
     explicit Material();
 
     /// Resource ID of the material.
-    uint64_t id;
+    uint32_t id;
 
     /// The resource ID of the shader to use for this material.
-    uint64_t shader;
+    uint32_t shader;
 
     /// Mapping of texture "sampler" names (available to the shader) to resouce
     /// IDs.
-    std::map<std::string, uint64_t> textures;
+    std::map<std::string, uint32_t> textures;
 
     /// If the shader adds light to the scene, the material is emissive.
     bool emissive;
