+++ src/shared/types/weak_hit.hpp
@@ -7,14 +7,14 @@
 namespace fr {
 
 struct WeakHit {
-    explicit WeakHit(uint64_t worker);
+    explicit WeakHit(uint32_t worker);
 
-    explicit WeakHit(uint64_t worker, float t);
+    explicit WeakHit(uint32_t worker, float t);
 
     explicit WeakHit();
 
     /// Resource ID of worker this hit occurred on.
-    uint64_t worker;
+    uint32_t worker;
 
     /// The t value of the intersection.
     float t;
